+++ src/pages/MyInformationPage.tsx
@@ -1,28 +1,19 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { Avatar, Box, Container, Link, Tab, Tabs, Typography, Button } from '@mui/material';
+import { Avatar, Box, Container, Tab, Tabs, Typography, Button } from '@mui/material';
 import { useAuth } from '../AuthenticationContext';
 import { MyInformation } from '../types/myInformation';
 import getMyInformation from '../api/myInformation';
-import AlbumReviewSummary from '../components/review/AlbumReviewSummary';
-import AlbumCover from '../components/album/AlbumCover';
 import ChangeName from '../components/common/Modal/ChangeName';
-import { ReactComponent as Hamburger } from '../assets/icons/hamburger.svg';
-import { ReactComponent as Edit } from '../assets/icons/edit-review.svg';
-import { ReactComponent as Delete } from '../assets/icons/delete-review.svg';
-import { StyledMenu, StyledMenuItem } from '../components/common/StyledMenu';
-import BASE_URL from '../config';
-import { typography } from '../constants/themeValue';
 import TabPanel from '../components/common/Tabs/TabPanel';
 import useTabs from '../hooks/useTab';
+import MyReviews from '../components/myInformation/MyReviews';
 
 function MyPage() {
   const [data, setData] = useState<MyInformation | null>(null);
   const navigate = useNavigate();
   const [user] = useAuth();
-  const [openMenu, setOpenMenu] = useState<(EventTarget & HTMLDivElement) | null>(null);
-  const [toEditReview, setToEditReview] = useState<string | null>(null);
   const { currentTab, handleChangeCurrentTab, tabProps } = useTabs('my-information-tab');
   const [open, setOpen] = useState(false);
   useEffect(() => {
@@ -31,6 +22,7 @@ function MyPage() {
 
       if (res.success) {
         setData(res.data);
+        console.log(res.data);
       }
     })();
   }, []);
@@ -38,12 +30,6 @@ function MyPage() {
   if (!user) {
     navigate(`/login`);
   }
-
-  const deleteMyReview = async (id: string) => {
-    await fetch(`${BASE_URL}/album/api/review/delete/${id}`, {
-      method: 'DELETE',
-    });
-  };
 
   return (
     <Container>
@@ -68,72 +54,7 @@ function MyPage() {
       </Tabs>
 
       <TabPanel value={currentTab} index={0}>
-        {data?.reviews && (
-          <Box sx={{ display: 'flex', flexWrap: 'wrap', mt: 2 }} gap={3}>
-            {data?.reviews.map((review) => (
-              <Box
-                sx={{
-                  position: 'relative',
-                  border: '1px solid rgb(52, 52, 52)',
-                  padding: '16px',
-                  borderRadius: '0px 16px 16px 16px',
-                  width: '282px',
-                  minWidth: '282px',
-                }}
-              >
-                <Box
-                  sx={{
-                    position: 'absolute',
-                    zIndex: 10,
-                    width: '24px',
-                    height: '24px',
-                    borderRadius: '50%',
-                    top: '27px',
-                    right: '27px',
-                    ':hover': { backgroundColor: 'rgb(126, 126, 126)' },
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                  }}
-                  onClick={(e) => {
-                    setToEditReview(review._id);
-                    setOpenMenu(e.currentTarget);
-                  }}
-                >
-                  <Hamburger>열기</Hamburger>
-                </Box>
-                <StyledMenu width={100} anchorEl={openMenu} open={Boolean(openMenu)} onClose={() => setOpenMenu(null)}>
-                  <Link sx={{ textDecoration: 'none' }} href={`/album/review/edit/${toEditReview}`}>
-                    <StyledMenuItem sx={{ height: '30px', padding: '9.75px' }}>
-                      <Edit />
-                      <Typography fontSize={typography.size.md} ml={2}>
-                        수정
-                      </Typography>
-                    </StyledMenuItem>
-                  </Link>
-                  <StyledMenuItem
-                    sx={{ height: '30px', padding: '9.75px' }}
-                    onClick={() => {
-                      deleteMyReview(toEditReview!);
-                    }}
-                  >
-                    <Delete />
-                    <Typography fontSize={typography.size.md} ml={2}>
-                      삭제
-                    </Typography>
-                  </StyledMenuItem>
-                </StyledMenu>
-                <AlbumCover
-                  albumId={review.albumId}
-                  url={review.thumbnail}
-                  albumTitle={review.albumName}
-                  artists={review.artists}
-                />
-                <AlbumReviewSummary review={review} isMyReview />
-              </Box>
-            ))}
-          </Box>
-        )}
+        {data?.reviews && <MyReviews reviews={data?.reviews} />}
       </TabPanel>
     </Container>
   );
